Remove warnings

+++ Handmade/Design.cpp
@@ -228,12 +228,12 @@ State* Design::Update(int deltaTime)
 	auto dimension = m_sceneCamera->GetResolution();
 
 	sceneBox.SetPosition(dimension.x * 0.5f, dimension.y * 0.5f, 0.0f);
-	sceneBox.SetDimension(dimension.x, dimension.y, 0.0f);
+	sceneBox.SetDimension(static_cast<GLfloat>(dimension.x), static_cast<GLfloat>(dimension.y), 0.0f);
 	sceneBox.Update();
 
 	BoxCollider mouseBox;
 	auto mousePosition = Input::Instance()->GetMousePosition();
-	mouseBox.SetPosition(mousePosition.x, mousePosition.y, 0.0f);
+	mouseBox.SetPosition(static_cast<GLfloat>(mousePosition.x), static_cast<GLfloat>(mousePosition.y), 0.0f);
 	mouseBox.SetDimension(1.0f, 1.0f, 0.0f);
 	mouseBox.Update();
